% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_classes}
\alias{yeast_classes}
\title{Classes for the yeast data}
\format{An object of class \code{integer} of length 10.}
\usage{
yeast_classes
}
\description{
See the \link{yeast} help page for further info
}
\keyword{datasets}
